reorganize packages locations and levels
the processing package can be used for both metadata and data,
so this pacakge is now directly under the client package

the pipeline, and its children (crypto and storage) are however only used for datastorage,
and thus it is more logical for those to be under the datastor package
