+++ client/datastor/pipeline/async_splitter.go
@@ -23,8 +23,8 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/zero-os/0-stor/client/datastor/pipeline/storage"
 	"github.com/zero-os/0-stor/client/metastor"
-	"github.com/zero-os/0-stor/client/pipeline/storage"
 
 	"golang.org/x/sync/errgroup"
 )
