+++ client/datastor/pipeline/config.go
@@ -21,9 +21,9 @@ import (
 	"runtime"
 
 	"github.com/zero-os/0-stor/client/datastor"
-	"github.com/zero-os/0-stor/client/pipeline/crypto"
-	"github.com/zero-os/0-stor/client/pipeline/processing"
-	"github.com/zero-os/0-stor/client/pipeline/storage"
+	"github.com/zero-os/0-stor/client/datastor/pipeline/crypto"
+	"github.com/zero-os/0-stor/client/datastor/pipeline/storage"
+	"github.com/zero-os/0-stor/client/processing"
 )
 
 var (
