+++ client/datastor/pipeline/crypto/doc.go
