+++ client/datastor/pipeline/crypto/hash.go
