+++ client/datastor/pipeline/crypto/hash_test.go
