+++ client/datastor/pipeline/crypto/sha.go
