+++ client/datastor/pipeline/crypto/sha_test.go
