+++ client/datastor/pipeline/single_object.go
@@ -22,8 +22,8 @@ import (
 	"io"
 	"io/ioutil"
 
+	"github.com/zero-os/0-stor/client/datastor/pipeline/storage"
 	"github.com/zero-os/0-stor/client/metastor"
-	"github.com/zero-os/0-stor/client/pipeline/storage"
 )
 
 // NewSingleObjectPipeline creates single-threaded pipeline
