+++ client/datastor/pipeline/single_object_test.go
@@ -19,8 +19,8 @@ package pipeline
 import (
 	"testing"
 
-	"github.com/zero-os/0-stor/client/pipeline/crypto"
-	"github.com/zero-os/0-stor/client/pipeline/processing"
+	"github.com/zero-os/0-stor/client/datastor/pipeline/crypto"
+	"github.com/zero-os/0-stor/client/processing"
 
 	"github.com/stretchr/testify/require"
 )
