+++ client/datastor/pipeline/storage/random.go
