+++ client/datastor/pipeline/storage/replicated.go
