+++ client/datastor/pipeline/storage/replicated_test.go
