+++ client/processing/compress.go
