+++ client/processing/compress_test.go
