+++ client/processing/compression_mode.go
