+++ client/processing/compression_mode_test.go
