+++ client/processing/compression_type.go
