+++ client/processing/compression_type_test.go
