+++ client/processing/crypto.go
