+++ client/processing/crypto_test.go
