+++ client/processing/encryption_type.go
