+++ client/processing/encryption_type_test.go
