+++ client/processing/processor.go
