+++ client/processing/processor_test.go
