+++ daemon/api/grpc/utils.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/zero-os/0-stor/client/datastor"
 
+	"github.com/zero-os/0-stor/client/datastor/pipeline/storage"
 	"github.com/zero-os/0-stor/client/metastor"
-	"github.com/zero-os/0-stor/client/pipeline/storage"
 	"github.com/zero-os/0-stor/daemon/api/grpc/rpctypes"
 	pb "github.com/zero-os/0-stor/daemon/api/grpc/schema"
 )
