+++ examples/hello_world/main.go
@@ -21,8 +21,8 @@ import (
 	"log"
 
 	"github.com/zero-os/0-stor/client"
-	"github.com/zero-os/0-stor/client/pipeline"
-	"github.com/zero-os/0-stor/client/pipeline/processing"
+	"github.com/zero-os/0-stor/client/datastor/pipeline"
+	"github.com/zero-os/0-stor/client/processing"
 )
 
 func main() {
